+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 - Fix sorting in borehole tables in viewer and editor mode.
 - Fix text displayed in backfill tab, when no backfill is available.
 - Fix text displayed in instrument tab, when no instrument is available.
+- Fix dialog message when enabling user.